+++ Dumper/Dumper/include/common.hpp
@@ -1,5 +1,5 @@
 #pragma once
 #include <YYToolkit/YYTK_Shared.hpp>
 
-static const char* const VERSION = "0.1.0";
 static YYTK::YYTKInterface* g_ModuleInterface = nullptr;
+static const char* const VERSION = "0.2.0";