+++ Dumper/CHANGELOG.md
@@ -1,6 +1,48 @@
 # Changelog
 
+https://keepachangelog.com/en/1.1.0/
+
 ## [Unreleased]
+
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## v0.2.0 - 2025-08-19
+
+### Added
+
+- Add `DumpHookVariables` that automatically dumps all arguments of a hook (`Self`, `Other`,
+  `Result`, `Arguments`).
+- Add `RegisterHook` that abstracts away the hook creation.
+- Add `RegisterHooks` that create multiple hooks with just one call.
+- Add `EnableDumper`.
+- Add `DispbleDumper`.
+
+### Changed
+
+- Add `dont_queue` parameter to `ToJsonObject`. When `true`, any struct found during the call
+  will not be added to the `queue`.
+- Export public interface functions to DLL. Not all functions has calling helper.
+- All dumping related functions now obey `g_SHOULD_DUMP`. If it is `false`, dumping will be skipped.
+
+### Deprecated
+### Removed
+
+- (BREAKING CHANGES) Overload of `DumpInstance`. All functions with default index filename, "index.html" now has suffix
+  `WithDefaultIndexFilename`. E.g. `DumpRValueWithDefaultIndexFilename`. The `CInstance` version
+  is also being explicitly named. E.g. `DumpCInstance`.
+
+### Fixed
+
+- The dumper does not dump the initial instance.
+- The `visited_pointers` was not cleared when the `DumpInstance` is called multiple times
+  during a session.
+
+### Security
 
 ## v0.1.0 - 2025-08-18